Refactor expense calculation logic, update product price column to float, and remove redundant logo from base template.

+++ sql_model/entities.py
@@ -61,7 +61,7 @@ class Product(Base):
     
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
     name: Mapped[str] = mapped_column(String, unique=True, nullable=False)
-    price: Mapped[float] = mapped_column(Integer, nullable=False)
+    price: Mapped[float] = mapped_column(Float, nullable=False)
     
     # Relationships
     sales: Mapped[List["Sale"]] = relationship("Sale", back_populates="product")
